# Array and guess